+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Compilation error regarding `Cronex.Table` genserver process
+- `Cronex.Table` ping message handling
 
 ## Version 0.1.0 - 2016/11/02